Moving the learner packages from the contributed experiments project into drools-core
I fixed as many compilation errors as I could although we need to look back because there are still errors about
using Drools 5.x methods that are incompatible with 6.x.

Also, there is one class called org.drools.learner.builder.BoostedTester that is missing and I'm not sure what
that class was supposed to do yet.